Add files via upload
A friend made a small machine with meccano to rotate a jam jar full of sand. This was to polish small brass pieces the he used in clock and model making. It had a handle to turn the jar through gearing. It became tedious to wind the handle so in version 2 he added a motor which was powered by an adjustable voltage power block. But even at the lowest voltage the speed was way too high, and the jar nearly bounced off the machine! :-) So I thought I'd have a go at designing a speed control based on the L9110 H-Bridge module controlled from an Arduino nano. Also just to make it more fun I would add a coloured display, to give speed, direction and auto run options. I used a rotary encoder to control the selections on the display. Check out the video.  YouTube/@Bathog44